Added additional permission checks for Greeting (recording) upload and download

+++ app/voicemail_greetings/voicemail_greetings.php
@@ -231,12 +231,16 @@ else {
 	echo "		<td align='left' nowrap>\n";
 	echo "			<span class='title'>".$text['title']."</span><br>\n";
 	echo "		</td>";
-	echo "		<td align='right' nowrap>\n";
-	echo "			<input type='button' class='btn' name='' alt='back' onclick=\"javascript:history.back();\" value='".$text['button-back']."'>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;\n";
-	echo "			<input name=\"file\" type=\"file\" class=\"formfld\" id=\"file\">\n";
-	echo "			<input name=\"type\" type=\"hidden\" value=\"rec\">\n";
-	echo "			<input name=\"submit\" type=\"submit\" class=\"btn\" id=\"upload\" value=\"".$text['button-save']."\">\n";
-	echo "		</td>\n";
+
+	if (permission_exists('recording_upload')) {
+		echo "	<td align='right' nowrap>\n";
+		echo "		<input type='button' class='btn' name='' alt='back' onclick=\"javascript:history.back();\" value='".$text['button-back']."'>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;\n";
+		echo "		<input name=\"file\" type=\"file\" class=\"formfld\" id=\"file\">\n";
+		echo "		<input name=\"type\" type=\"hidden\" value=\"rec\">\n";
+		echo "		<input name=\"submit\" type=\"submit\" class=\"btn\" id=\"upload\" value=\"".$text['button-save']."\">\n";
+		echo "	</td>\n";
+	}
+
 	echo "	</tr>";
 	echo "	<tr>";
 	echo "		<td align='left' colspan='2'>\n";
@@ -317,12 +321,16 @@ else {
 			echo "	</td>\n";
 			echo "	<td valign='top' class='".$row_style[$c]."'>".$row['greeting_name']."</td>\n";
 			echo "	<td valign='top' class='".$row_style[$c]."'>";
-			echo "		<a href=\"voicemail_greetings.php?id=$voicemail_id&a=download&type=rec&t=bin&filename=".base64_encode($row['greeting_name'])."\">".$text['table-download']."</a>";
-			//echo "		&nbsp;\n";
-			//echo "		<a href=\"javascript:void(0);\" onclick=\"window.open('voicemail_greeting_play.php?id=$voicemail_id&a=download&type=rec&filename=".base64_encode($row['greeting_name'])."', 'play',' width=420,height=40,menubar=no,status=no,toolbar=no')\">\n";
-			//echo "		play";
-			//echo "		</a>";
-			echo 	"</td>\n";
+
+			if (permission_exists('recording_download')) {
+				echo "	<a href=\"voicemail_greetings.php?id=$voicemail_id&a=download&type=rec&t=bin&filename=".base64_encode($row['greeting_name'])."\">".$text['table-download']."</a>";
+			}
+				//echo "&nbsp;\n";
+				//echo "<a href=\"javascript:void(0);\" onclick=\"window.open('voicemail_greeting_play.php?id=$voicemail_id&a=download&type=rec&filename=".base64_encode($row['greeting_name'])."', 'play',' width=420,height=40,menubar=no,status=no,toolbar=no')\">\n";
+				//echo "play";
+				//echo "</a>";
+
+			echo "	</td>\n";
 			echo "	<td class='".$row_style[$c]."' nowrap>".$tmp_filesize."</td>\n";
 			echo "	<td valign='top' class='row_stylebg'>".$row['greeting_description']."&nbsp;</td>\n";
 			echo "	<td valign='top' align='right'>\n";
